[Delete] MEMEAuthAPI 삭제, MEMEAPI와 통합
