+++ MEME/MEME/Core/Model/SocialProvider.swift
@@ -1,0 +1,13 @@
+//
+//  SocialProvider.swift
+//  MEME
+//
+//  Created by 이동현 on 4/7/24.
+//
+
+import Foundation
+
+enum SocialProvider: String {
+    case APPLE
+    case KAKAO
+}
