+++ MEME/MEME/Login/Login/View/LoginViewController.swift
@@ -1,0 +1,19 @@
+//
+//  LoginViewController.swift
+//  MEME
+//
+//  Created by 이동현 on 1/9/24.
+//
+
+import UIKit
+import KakaoSDKUser
+import AuthenticationServices
+
+final class LoginViewController: UIViewController {
+    private let userDefaultManager = UserDefaultManager.shared
+    private let authManager = AuthManager.shared
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
